Added a ValueSlider that subclasses the FloatRangeSlider widget. Rather than percentages, it represents the full range of possible values for the corresponding quantity. It includes a multiplier field that allows the displayed values to be scaled versions of the "real" values to which they correspond.
The SpectrogramWidget was modified to use the slice functionality
of Spectrogram to avoid recomputation.
On clicking within the spectrogram, the spectrum at that time is
computed, but the range displayed on the left axes goes from the mean
value to the maximum. It should probably have vertical gridlines, but
I haven't done that yet.

I added a checkbox to show the raw voltages as a semitransparent overlay
at the top of the spectrogram.