Fixed a problem with thumbnail in DigFile which was computing times incorrectly when the start time was nonzero.
Also tweaked the __str__ method to comma-separate the number of values.

Modified the __str__ method of Spectrogram to include the dimensions
of the intensity array. I have also changed the default overlap
from 1/8 (meaning that 1/8th of the points are reused between adjacent
spectra in the spectrogram) to 3/4, which makes for smoother but much
larger spectra. This parameter may be overridden by explicitly
passing overlap_shift_factor=1/8, for example.