A modest clean-up of this simple follower to make explicit the parameters used to handle discontinuities.
smoothing: describes the number of values on either side of a point to average over to produce the smoothed version of
intensities to use in looking for a peak.

max_hop: the largest number of pixels acceptable between peaks at successive time steps.
If the jump exceeds this value, following terminates.